Add support for re-queuing half-processed ops to honour yield()
For ops that may contain many objects or events, such as OP_DR,
a yield() call would previously have no affect on these events
but would continue to get processed until all events of the op
were exhausted.

With this fix the various op serving call paths will now check for yield
and re-enqueue the op for the next call to poll() or similar.